+++ osu.Desktop.VisualTests/Tests/TestCaseDrawableRoom.cs
@@ -17,7 +17,7 @@ namespace osu.Desktop.VisualTests.Tests
     {
         public override string Description => @"Select your favourite room";
 
-        private RulesetDatabase rulesets;
+        private RulesetStore rulesets;
 
         protected override void LoadComplete()
         {
@@ -125,7 +125,7 @@ namespace osu.Desktop.VisualTests.Tests
         }
 
         [BackgroundDependencyLoader]
-        private void load(RulesetDatabase rulesets)
+        private void load(RulesetStore rulesets)
         {
             this.rulesets = rulesets;
         }
