+++ osu.Desktop.VisualTests/Tests/TestCaseRoomInspector.cs
@@ -16,7 +16,7 @@ namespace osu.Desktop.VisualTests.Tests
     {
         public override string Description => @"from the multiplayer lobby";
 
-        private RulesetDatabase rulesets;
+        private RulesetStore rulesets;
 
         protected override void LoadComplete()
         {
@@ -136,7 +136,7 @@ namespace osu.Desktop.VisualTests.Tests
         }
 
         [BackgroundDependencyLoader]
-        private void load(RulesetDatabase rulesets)
+        private void load(RulesetStore rulesets)
         {
             this.rulesets = rulesets;
         }
