+++ osu.Game/IPC/ScoreIPCChannel.cs
@@ -10,9 +10,9 @@ namespace osu.Game.IPC
 {
     public class ScoreIPCChannel : IpcChannel<ScoreImportMessage>
     {
-        private readonly ScoreDatabase scores;
+        private readonly ScoreStore scores;
 
-        public ScoreIPCChannel(IIpcHost host, ScoreDatabase scores = null)
+        public ScoreIPCChannel(IIpcHost host, ScoreStore scores = null)
             : base(host)
         {
             this.scores = scores;
