+++ osu.Game/Overlays/Settings/Sections/GameplaySection.cs
@@ -25,7 +25,7 @@ namespace osu.Game.Overlays.Settings.Sections
         }
 
         [BackgroundDependencyLoader]
-        private void load(RulesetDatabase rulesets)
+        private void load(RulesetStore rulesets)
         {
             foreach(Ruleset ruleset in rulesets.AllRulesets.Select(info => info.CreateInstance()))
             {
